Main function + working make file
This is the first correct update to the files. The code now uses MinGW as a compiler program, and uses the included makefile to make the compiling process easier for the user. The main.c file does not have any useful code as of yet but will be the foundation to our mechanical tester. The makefile has not been adapted to run multiple libraries at once yet (i believe) however this will also come in future updates when libraries of the hardware will be used.